license changes
