+++ lib/kcerts/BUILD.bazel
@@ -14,9 +14,9 @@ go_library(
     deps = [
         "//lib/cache:go_default_library",
         "//lib/config/marshal:go_default_library",
+        "//lib/kcerts/ked25519:go_default_library",
         "//lib/logger:go_default_library",
         "@com_github_mitchellh_go_homedir//:go_default_library",
-        "@org_golang_x_crypto//ed25519:go_default_library",
         "@org_golang_x_crypto//ssh:go_default_library",
         "@org_golang_x_crypto//ssh/agent:go_default_library",
     ],
