+++ lib/kcerts/keys.go
@@ -7,6 +7,7 @@ import (
 	"crypto/rsa"
 	"crypto/x509"
 	"encoding/pem"
+	"github.com/enfabrica/enkit/lib/kcerts/ked25519"
 	"golang.org/x/crypto/ssh"
 )
 
@@ -36,7 +37,7 @@ func (e ed25519Provider) Signer() crypto.Signer {
 }
 
 func (e ed25519Provider) SSHPemEncode() ([]byte, error) {
-	res, err := OpenSSHEncode21559PrivateKey(e.rawKey)
+	res, err := ked25519.OpenSSHEncode21559PrivateKey(e.rawKey)
 	if err != nil {
 		return nil, err
 	}
